+++ app/src/main/java/com/c7z/myapplication/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.fragment.app.Fragment
 import com.c7z.myapplication.databinding.ActivityMainBinding
 import com.c7z.myapplication.group.GroupFragment
 import com.c7z.myapplication.home.HomeFragment
-import com.c7z.myapplication.map.MapFragment
+import com.c7z.myapplication.map.NaverMapFragment
 import com.c7z.myapplication.user.UserFragment
 import com.c7z.myapplication.todo.TodoFragment
 
@@ -34,7 +34,7 @@ class MainActivity : AppCompatActivity() {
             when(it.itemId) {
                 R.id.main_bottom_nav_btn_home -> setFragment(HomeFragment())
                 R.id.main_bottom_nav_btn_group -> setFragment(GroupFragment())
-                R.id.main_bottom_nav_btn_map -> setFragment(MapFragment())
+                R.id.main_bottom_nav_btn_map -> setFragment(NaverMapFragment())
                 R.id.main_bottom_nav_btn_todo -> setFragment(TodoFragment())
                 R.id.main_bottom_nav_btn_user -> setFragment(UserFragment())
             }
