--- app/src/main/java/com/c7z/myapplication/map/MapFragment.kt
@@ -1,21 +1,0 @@
-package com.c7z.myapplication.map
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.c7z.myapplication.databinding.FragmentMapBinding
-
-class MapFragment : Fragment() {
-    private lateinit var binding : FragmentMapBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentMapBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-}
